+++ src/compnents/organisms/BookDetailCard.tsx
@@ -1,0 +1,66 @@
+import { Image } from "@chakra-ui/image";
+import { Box, Stack, Text } from "@chakra-ui/layout";
+import { VFC } from "react";
+import { SecondaryButton } from "../atoms/SecondaryButton";
+
+type BookType = {
+  title: string;
+  article: string;
+  description: string;
+  image: string;
+};
+
+type Props = {
+  isLoading: boolean;
+  book: BookType;
+  onClick: () => void;
+};
+
+export const BookDetailCard: VFC<Props> = ({ book, isLoading, onClick }) => {
+  return (
+    <Box
+      w="560px"
+      h="630px"
+      bg="gray.100"
+      borderRadius="10px"
+      shadow="md"
+      overflow="hidden"
+      mx="auto"
+      mt={10}
+      p={2}
+    >
+      <Image
+        height="300px"
+        width="360px"
+        objectFit="cover"
+        src={book.image}
+        mx="auto"
+      />
+      <Stack spacing={6} mt={5} p={4}>
+        <Box>
+          <Text as="h2">タイトル</Text>
+          <Text fontSize="sm" pl={4} pt={2}>
+            {book.title}
+          </Text>
+        </Box>
+        <Box>
+          <Text as="h2">著者名</Text>
+          <Text fontSize="sm" pl={4} pt={2}>
+            {book.article}
+          </Text>
+        </Box>
+        <Box>
+          <Text as="h2">説明</Text>
+          <Text fontSize="sm" pl={4} pt={2}>
+            {book.description}
+          </Text>
+        </Box>
+      </Stack>
+      <Box textAlign="center">
+        <SecondaryButton onClick={onClick} isLoading={isLoading}>
+          削除
+        </SecondaryButton>
+      </Box>
+    </Box>
+  );
+};
